+++ mud-action/src/main/java/com/jpinfo/mudengine/action/repository/MudActionRepository.java
@@ -14,10 +14,10 @@ public interface MudActionRepository extends CrudRepository<MudAction, Long> {
 			+ "and not exists (select 1 from MudAction o where o.actorCode=a.actorCode and o.currState=1)")
 	List<MudAction> findStartableActions();
 	
-	List<MudAction> findByIssuerCode(Integer issuerCode);
+	List<MudAction> findByIssuerCode(Long issuerCode);
 	
 	@Query("select a from MudAction a where a.currState=1 and a.endTurn <= ?")
 	List<MudAction> findFinishedActions(Long curTurn);
 	
-	MudAction findFirstOneByCurrStateAndActorCode(Integer currState, Integer actorCode);
+	MudAction findFirstOneByCurrStateAndActorCode(Integer currState, Long actorCode);
 }
